refactor yolo eval scripts and remove yolo_utils to use common_demo_utils
